fix(nx-generators): align, refactor and format generators
