+++ packages/nx/src/generators/app-resources/files/__name__/iOS/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,122 +1,122 @@
 {
-  "images" : [
+  "images": [
     {
-      "size" : "20x20",
-      "idiom" : "iphone",
-      "filename" : "icon-20@2x.png",
-      "scale" : "2x"
+      "size": "20x20",
+      "idiom": "iphone",
+      "filename": "icon-20@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "20x20",
-      "idiom" : "iphone",
-      "filename" : "icon-20@3x.png",
-      "scale" : "3x"
+      "size": "20x20",
+      "idiom": "iphone",
+      "filename": "icon-20@3x.png",
+      "scale": "3x"
     },
     {
-      "size" : "29x29",
-      "idiom" : "iphone",
-      "filename" : "icon-29.png",
-      "scale" : "1x"
+      "size": "29x29",
+      "idiom": "iphone",
+      "filename": "icon-29.png",
+      "scale": "1x"
     },
     {
-      "size" : "29x29",
-      "idiom" : "iphone",
-      "filename" : "icon-29@2x.png",
-      "scale" : "2x"
+      "size": "29x29",
+      "idiom": "iphone",
+      "filename": "icon-29@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "29x29",
-      "idiom" : "iphone",
-      "filename" : "icon-29@3x.png",
-      "scale" : "3x"
+      "size": "29x29",
+      "idiom": "iphone",
+      "filename": "icon-29@3x.png",
+      "scale": "3x"
     },
     {
-      "size" : "40x40",
-      "idiom" : "iphone",
-      "filename" : "icon-40@2x.png",
-      "scale" : "2x"
+      "size": "40x40",
+      "idiom": "iphone",
+      "filename": "icon-40@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "40x40",
-      "idiom" : "iphone",
-      "filename" : "icon-40@3x.png",
-      "scale" : "3x"
+      "size": "40x40",
+      "idiom": "iphone",
+      "filename": "icon-40@3x.png",
+      "scale": "3x"
     },
     {
-      "size" : "60x60",
-      "idiom" : "iphone",
-      "filename" : "icon-60@2x.png",
-      "scale" : "2x"
+      "size": "60x60",
+      "idiom": "iphone",
+      "filename": "icon-60@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "60x60",
-      "idiom" : "iphone",
-      "filename" : "icon-60@3x.png",
-      "scale" : "3x"
+      "size": "60x60",
+      "idiom": "iphone",
+      "filename": "icon-60@3x.png",
+      "scale": "3x"
     },
     {
-      "size" : "20x20",
-      "idiom" : "ipad",
-      "filename" : "icon-20.png",
-      "scale" : "1x"
+      "size": "20x20",
+      "idiom": "ipad",
+      "filename": "icon-20.png",
+      "scale": "1x"
     },
     {
-      "size" : "20x20",
-      "idiom" : "ipad",
-      "filename" : "icon-20@2x.png",
-      "scale" : "2x"
+      "size": "20x20",
+      "idiom": "ipad",
+      "filename": "icon-20@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "29x29",
-      "idiom" : "ipad",
-      "filename" : "icon-29.png",
-      "scale" : "1x"
+      "size": "29x29",
+      "idiom": "ipad",
+      "filename": "icon-29.png",
+      "scale": "1x"
     },
     {
-      "size" : "29x29",
-      "idiom" : "ipad",
-      "filename" : "icon-29@2x.png",
-      "scale" : "2x"
+      "size": "29x29",
+      "idiom": "ipad",
+      "filename": "icon-29@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "40x40",
-      "idiom" : "ipad",
-      "filename" : "icon-40.png",
-      "scale" : "1x"
+      "size": "40x40",
+      "idiom": "ipad",
+      "filename": "icon-40.png",
+      "scale": "1x"
     },
     {
-      "size" : "40x40",
-      "idiom" : "ipad",
-      "filename" : "icon-40@2x.png",
-      "scale" : "2x"
+      "size": "40x40",
+      "idiom": "ipad",
+      "filename": "icon-40@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "76x76",
-      "idiom" : "ipad",
-      "filename" : "icon-76.png",
-      "scale" : "1x"
+      "size": "76x76",
+      "idiom": "ipad",
+      "filename": "icon-76.png",
+      "scale": "1x"
     },
     {
-      "size" : "76x76",
-      "idiom" : "ipad",
-      "filename" : "icon-76@2x.png",
-      "scale" : "2x"
+      "size": "76x76",
+      "idiom": "ipad",
+      "filename": "icon-76@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "83.5x83.5",
-      "idiom" : "ipad",
-      "filename" : "icon-83.5@2x.png",
-      "scale" : "2x"
+      "size": "83.5x83.5",
+      "idiom": "ipad",
+      "filename": "icon-83.5@2x.png",
+      "scale": "2x"
     },
     {
-      "size" : "1024x1024",
-      "idiom" : "ios-marketing",
-      "filename" : "icon-1024.png",
-      "scale" : "1x"
+      "size": "1024x1024",
+      "idiom": "ios-marketing",
+      "filename": "icon-1024.png",
+      "scale": "1x"
     }
   ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
+  "info": {
+    "version": 1,
+    "author": "xcode"
   }
 }
