+++ packages/nx/src/generators/app-resources/files/__name__/iOS/Assets.xcassets/LaunchScreen.Center.imageset/Contents.json
@@ -1,23 +1,23 @@
 {
-  "images" : [
+  "images": [
     {
-      "idiom" : "universal",
-      "filename" : "LaunchScreen-Center.png",
-      "scale" : "1x"
+      "idiom": "universal",
+      "filename": "LaunchScreen-Center.png",
+      "scale": "1x"
     },
     {
-      "idiom" : "universal",
-      "filename" : "LaunchScreen-Center@2x.png",
-      "scale" : "2x"
+      "idiom": "universal",
+      "filename": "LaunchScreen-Center@2x.png",
+      "scale": "2x"
     },
     {
-      "idiom" : "universal",
-      "filename" : "LaunchScreen-Center@3x.png",
-      "scale" : "3x"
+      "idiom": "universal",
+      "filename": "LaunchScreen-Center@3x.png",
+      "scale": "3x"
     }
   ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
+  "info": {
+    "version": 1,
+    "author": "xcode"
   }
 }
