+++ packages/nx/src/generators/app-resources/schema.json
@@ -14,7 +14,5 @@
       "default": "App_Resources"
     }
   },
-  "required": [
-    "path"
-  ]
+  "required": ["path"]
 }
