+++ packages/nx/src/generators/application/files_angular/__dot__eslintrc.json
@@ -1,1 +1,5 @@
-{ "extends": "<%= projectRootOffset %>.eslintrc.json", "ignorePatterns": ["!**/*", "references.d.ts", "node_modules/**/*", "hooks/**/*", "platforms/**/*"], "rules": {} }
+{
+  "extends": "<%= projectRootOffset %>.eslintrc.json",
+  "ignorePatterns": ["!**/*", "references.d.ts", "node_modules/**/*", "hooks/**/*", "platforms/**/*"],
+  "rules": {}
+}
