+++ packages/nx/src/generators/application/files_angular/src/app.component.html
@@ -1,13 +1,3 @@
-<% if (routing) { %><GridLayout>
-  <page-router-outlet></page-router-outlet>
-</GridLayout><% }
-else { %>
-<ActionBar title="My App" class="action-bar">
-</ActionBar>
-<StackLayout class="page p-20">
-  <Label
-    text="Hello"
-    class="h1 text-center c-white"
-  ></Label>
-</StackLayout
-><% } %>
+<% if (routing) { %><GridLayout> <page-router-outlet></page-router-outlet> </GridLayout><% } else { %>
+<ActionBar title="My App" class="action-bar"> </ActionBar>
+<StackLayout class="page p-20"> <Label text="Hello" class="h1 text-center c-white"></Label> </StackLayout><% } %>
