+++ packages/nx/src/generators/application/files_angular/src/app.ios.scss
