+++ packages/nx/src/generators/application/files_angular/src/scss/_index.scss
@@ -3,31 +3,30 @@
  * for specific overrides: see the overrides.{ios,android}.scss
  */
 
- @import 'variables';
- @import 'mixins';
- 
- // {N} core theme rulesets (mix/match these or only use what you need)
- @import 'nativescript-theme-core/scss/action-bar';
- @import 'nativescript-theme-core/scss/activity-indicator';
- @import 'nativescript-theme-core/scss/buttons';
- @import 'nativescript-theme-core/scss/icons';
- @import 'nativescript-theme-core/scss/forms';
- @import 'nativescript-theme-core/scss/labels';
- @import 'nativescript-theme-core/scss/list-view';
- @import 'nativescript-theme-core/scss/page';
- @import 'nativescript-theme-core/scss/progress';
- @import 'nativescript-theme-core/scss/segmented-bar';
- @import 'nativescript-theme-core/scss/side-drawer';
- @import 'nativescript-theme-core/scss/slider';
- @import 'nativescript-theme-core/scss/switch';
- @import 'nativescript-theme-core/scss/tabs';
- 
- // {N} custom styles (create/import other scss files or define as needed)
- // -- START custom --
- @import 'action-bar';
- @import 'tags';
- // -- END custom
- 
- // since {N} core theme utilities are additive to any above class, these should always be last
- @import 'nativescript-theme-core/scss/utilities/index';
- 
+@import 'variables';
+@import 'mixins';
+
+// {N} core theme rulesets (mix/match these or only use what you need)
+@import 'nativescript-theme-core/scss/action-bar';
+@import 'nativescript-theme-core/scss/activity-indicator';
+@import 'nativescript-theme-core/scss/buttons';
+@import 'nativescript-theme-core/scss/icons';
+@import 'nativescript-theme-core/scss/forms';
+@import 'nativescript-theme-core/scss/labels';
+@import 'nativescript-theme-core/scss/list-view';
+@import 'nativescript-theme-core/scss/page';
+@import 'nativescript-theme-core/scss/progress';
+@import 'nativescript-theme-core/scss/segmented-bar';
+@import 'nativescript-theme-core/scss/side-drawer';
+@import 'nativescript-theme-core/scss/slider';
+@import 'nativescript-theme-core/scss/switch';
+@import 'nativescript-theme-core/scss/tabs';
+
+// {N} custom styles (create/import other scss files or define as needed)
+// -- START custom --
+@import 'action-bar';
+@import 'tags';
+// -- END custom
+
+// since {N} core theme utilities are additive to any above class, these should always be last
+@import 'nativescript-theme-core/scss/utilities/index';
