+++ packages/nx/src/generators/application/files_angular/src/scss/_tags.scss
