+++ packages/nx/src/generators/application/files_angular_routing/src/features/home/components/home.component.html
@@ -1,5 +1,4 @@
-<ActionBar title="My App" class="action-bar" flat="true">
-</ActionBar>
+<ActionBar title="My App" class="action-bar" flat="true"> </ActionBar>
 <StackLayout class="p-20">
   <Label text="Enjoy" class="h1 text-center c-white"></Label>
   <Image src="~/assets/images/nx-nativescript.png" class="h-center m-t-20" width="300"></Image>
