+++ packages/nx/src/generators/application/files_vanilla/src/app.scss
