+++ packages/nx/src/generators/application/files_vanilla/webpack.config.js
@@ -1,8 +1,8 @@
-const webpack = require("@nativescript/webpack");
+const webpack = require('@nativescript/webpack');
 
 module.exports = (env) => {
-	webpack.init(env);
+  webpack.init(env);
   webpack.useConfig('typescript');
 
-	return webpack.resolveConfig();
+  return webpack.resolveConfig();
 };
