+++ packages/nx/src/generators/application/schema.d.ts
@@ -12,3 +12,16 @@ export interface ApplicationSchema {
   routing?: boolean;
   groupByName?: boolean;
 }
+
+export interface NormalizedSchema extends ApplicationSchema {
+  baseName: string;
+  projectName: string;
+  projectRoot: string;
+  projectSourceRoot: string;
+  projectRootOffset: string;
+  projectSourceRootOffset: string;
+  parsedTags: string[];
+  outputPath: string;
+  buildExecutor: '@nativescript/nx:build';
+  lintExecutor: '@nx/eslint:lint';
+}
