+++ packages/nx/src/generators/application/schema.json
@@ -17,7 +17,13 @@
         "$source": "argv",
         "index": 0
       },
+      "alias": "d",
       "x-prompt": "What directory would you like to use for the application? (full path - eg. apps/my-lib)"
+    },
+    "name": {
+      "type": "string",
+      "description": "The name of the application.",
+      "alias": "n"
     },
     "framework": {
       "description": "The frontend framework to use.",
@@ -37,11 +43,6 @@
           }
         ]
       }
-    },
-    "name": {
-      "type": "string",
-      "description": "The name of the application.",
-      "alias": "n"
     },
     "routing": {
       "type": "boolean",
