+++ packages/nx/src/generators/init/schema.d.ts
