+++ packages/nx/src/generators/library/schema.d.ts
@@ -16,3 +16,15 @@ export interface LibrarySchema {
   pascalCaseFiles?: boolean;
   js?: boolean;
 }
+
+export interface NormalizedSchema extends LibrarySchema {
+  baseName: string;
+  projectName: string;
+  projectRoot: string;
+  projectSourceRoot: string;
+  projectRootOffset: string;
+  projectSourceRootOffset: string;
+  parsedTags: string[];
+  outputPath: string;
+  lintExecutor: '@nx/eslint:lint';
+}
